comments + circular buffer index type fix

+++ include/utils/task_queue.h
@@ -7,12 +7,37 @@
 
 #define MAX_TASK_QUEUE_SIZE 20
 
+/**
+ * @brief Represents a single network-related task to be processed by a worker thread.
+ *
+ * @details This structure holds the:
+ * - accepted_peer_conn_t *connection: pointer to the accepted peer connection associated with this task.
+ * - int success_flag: success flag
+ * - int is_active: active/pending flag
+ */
 typedef struct {
   accepted_peer_conn_t *connection;
   int success_flag;
   int is_active;
 } network_task_t;
 
+
+/**
+ * @brief A thread pool managing a circular task queue and worker threads.
+ * 
+ * Implements SP-MC (single producer - multi consumer) pattern
+ *
+ * @details This structure holds the:
+ * pthread_t *workers: shared resource
+ * pthread_mutex_t mx_queue: ownership
+ * sem_t sem_empty_cnt: available spot count
+ * sem_t sem_fill_cnt: taken-up spot count
+ * network_task_t *tasks[MAX_TASK_QUEUE_SIZE]: bounded circular queue
+ * size_t head: insertion index
+ * size_t tail: extraction index
+ * int worker_cnt: number of resources allocated
+ * volatile int shutdown: immediate signal of thread pool termination
+ */
 typedef struct {
   // 8 byte aligned members
   pthread_t *workers;
@@ -24,8 +49,8 @@ typedef struct {
   network_task_t *tasks[MAX_TASK_QUEUE_SIZE];
 
   // 4 byte aligned members
-  uint head;
-  uint tail;
+  size_t head;
+  size_t tail;
   int worker_cnt;
   volatile int shutdown;
